- Promoted some document conventions to base conventions to reuse the LinqPathProvider on the file system query.
