+++ Raven.Client.Lightweight/Convention.cs
@@ -7,6 +7,8 @@ using System.Reflection;
 using System.Threading.Tasks;
 using Raven.Abstractions.Connection;
 using Raven.Client.Document;
+using Raven.Client.Linq;
+using System.Linq.Expressions;
 
 namespace Raven.Client
 {
@@ -102,5 +104,52 @@ namespace Raven.Client
 		/// in async manner
 		/// </summary>
 		public Func<HttpResponseMessage, OperationCredentials, Task<Action<HttpClient>>> HandleForbiddenResponseAsync { get; set; }
+
+
+        public delegate LinqPathProvider.Result CustomQueryTranslator(LinqPathProvider provider, Expression expression);
+
+        internal LinqPathProvider.Result TranslateCustomQueryExpression(LinqPathProvider provider, Expression expression)
+        {
+            var member = GetMemberInfoFromExpression(expression);
+
+            CustomQueryTranslator translator;
+            if (!customQueryTranslators.TryGetValue(member, out translator))
+                return null;
+
+            return translator.Invoke(provider, expression);
+        }
+
+        private static MemberInfo GetMemberInfoFromExpression(Expression expression)
+        {
+            var callExpression = expression as MethodCallExpression;
+            if (callExpression != null)
+                return callExpression.Method;
+
+            var memberExpression = expression as MemberExpression;
+            if (memberExpression != null)
+                return memberExpression.Member;
+
+            throw new NotSupportedException("A custom query translator can only be used to evaluate a simple member access or method call.");
+        }
+
+
+        private readonly Dictionary<MemberInfo, CustomQueryTranslator> customQueryTranslators = new Dictionary<MemberInfo, CustomQueryTranslator>();
+
+        public void RegisterCustomQueryTranslator<T>(Expression<Func<T, object>> member, CustomQueryTranslator translator)
+        {
+            var body = member.Body as UnaryExpression;
+            if (body == null)
+                throw new NotSupportedException("A custom query translator can only be used to evaluate a simple member access or method call.");
+
+            var info = GetMemberInfoFromExpression(body.Operand);
+
+            if (!customQueryTranslators.ContainsKey(info))
+                customQueryTranslators.Add(info, translator);
+        }
+
+        /// <summary>
+        /// Saves Enums as integers and instruct the Linq provider to query enums as integer values.
+        /// </summary>
+        public bool SaveEnumsAsIntegers { get; set; }
 	}
 }
